WLA-23: Move logic from Step to Declaration
